+++ src/main/java/blackjack/domain/player/Player.java
@@ -4,27 +4,57 @@ import blackjack.domain.card.Deck;
 import blackjack.domain.card.Hand;
 import java.util.List;
 
-public interface Player {
+public abstract class Player {
+    private final Name name;
+    private final Hand hand;
 
-    boolean isDrawable();
+    protected Player(final Name name, final Hand hand) {
+        this.name = name;
+        this.hand = hand;
+    }
 
-    boolean isDealer();
+    abstract public boolean isDrawable();
 
-    void initialDraw(Deck deck);
+    abstract public boolean isDealer();
 
-    void draw(final Deck deck);
+    public final void initialDraw(final Deck deck) {
+        draw(deck);
+        draw(deck);
+    }
 
-    int score();
+    public final void draw(final Deck deck) {
+        hand.add(deck.draw());
+    }
 
-    void stay();
+    public final int score() {
+        return hand.score();
+    }
 
-    Result play(final Hand hand);
+    public final void stay() {
+        hand.stay();
+    }
 
-    boolean isSameName(Name name);
+    public final Result play(final Hand hand) {
+        return this.hand.play(hand);
+    }
 
-    Name name();
+    public final boolean isSameName(final Name name) {
+        return this.name.equals(name);
+    }
 
-    String getNameValue();
+    public final Name name() {
+        return name;
+    }
 
-    List<String> getSymbols();
+    protected final Hand hand() {
+        return hand;
+    }
+
+    public final String getNameValue() {
+        return name.getValue();
+    }
+
+    public final List<String> getSymbols() {
+        return hand.getSymbols();
+    }
 }
